:hammer: extract AbstractGlobalCopywriter to commonMain for cross-platform reuse
Following #4335 (SupportedLanguages + LanguageCache in commonMain), each platform's
GlobalCopywriter still duplicated ~50 lines of identical boilerplate. Lift it into
a single abstract base class so subclasses only declare what actually differs.

AbstractGlobalCopywriter owns:
- LanguageCache wiring and validate-or-default-EN bootstrap
- protected `copywriter` mutableStateOf field
- switchLanguage with a protected `onLanguageSwitched(language)` hook for
  post-switch side effects
- getAllLanguages, language(), getText, getKeys, getDate, getAbridge

DesktopGlobalCopywriter now keeps only:
- TaskExecutor + TaskDao wiring
- onLanguageSwitched override that dispatches SwitchLanguageTask
- getText override implementing the desktop-only English fallback
  + bracketed-id missing-key UX

Net: -45 lines on desktop, +75 lines in commonMain that Android/iOS can adopt
in their own repos, replacing equivalent boilerplate there.

Closes #4337
