+++ app/src/desktopMain/kotlin/com/crosspaste/i18n/I18n.kt
@@ -1,15 +1,11 @@
 package com.crosspaste.i18n
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import com.crosspaste.config.CommonConfigManager
 import com.crosspaste.db.task.SwitchLanguageInfo
 import com.crosspaste.db.task.TaskDao
 import com.crosspaste.db.task.TaskType
 import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.EMPTY_STRING
 import com.crosspaste.i18n.SupportedLanguages.EN
-import com.crosspaste.i18n.SupportedLanguages.LANGUAGE_LIST
 import com.crosspaste.task.TaskExecutor
 import com.crosspaste.utils.DateTimeFormatOptions
 import com.crosspaste.utils.GlobalCoroutineScope.cpuCoroutineDispatcher
@@ -23,10 +19,10 @@ import java.nio.charset.StandardCharsets
 import java.util.Properties
 
 class DesktopGlobalCopywriter(
-    private val configManager: CommonConfigManager,
+    configManager: CommonConfigManager,
     private val lazyTaskExecutor: Lazy<TaskExecutor>,
     private val taskDao: TaskDao,
-) : GlobalCopywriter {
+) : AbstractGlobalCopywriter(configManager, { DesktopCopywriter(it) }) {
 
     companion object {
 
@@ -35,46 +31,17 @@ class DesktopGlobalCopywriter(
 
     private val logger = KotlinLogging.logger {}
 
-    private val cache = LanguageCache { DesktopCopywriter(it) }
-
-    private var language: String =
-        run {
-            val language = configManager.getCurrentConfig().language
-            if (!LANGUAGE_LIST.contains(language)) {
-                configManager.updateConfig("language", EN)
-                EN
-            } else {
-                language
-            }
-        }
-
-    // mutableStateOf is intentionally used here to trigger Compose recomposition when the
-    // language changes. Writes from non-Compose threads (e.g., switchLanguage()) are safe
-    // because Compose's global snapshot system handles cross-thread state mutations.
-    private var copywriter: Copywriter by mutableStateOf(cache.getOrCreate(language))
-
-    private val enCopywriter by lazy { cache.getOrCreate(EN) }
+    private val enCopywriter by lazy { copywriterFor(EN) }
 
     private val taskExecutor by lazy { lazyTaskExecutor.value }
 
-    override fun language(): String = copywriter.language()
-
-    override fun switchLanguage(language: String) {
-        logger.info { "Switching language $language" }
-        copywriter = cache.getOrCreate(language)
-        configManager.updateConfig("language", language)
+    override fun onLanguageSwitched(language: String) {
         cpuCoroutineDispatcher.launch {
             taskExecutor.submitTask(
                 taskDao.createTask(null, TaskType.SWITCH_LANGUAGE_TASK, SwitchLanguageInfo(language)),
             )
         }
     }
-
-    override fun getAllLanguages(): List<Language> =
-        LANGUAGE_LIST.map {
-            val cw = cache.getOrCreate(it)
-            Language(cw.getAbridge(), cw.getText("current_language"))
-        }
 
     override fun getText(
         id: String,
@@ -94,15 +61,6 @@ class DesktopGlobalCopywriter(
             text
         }
     }
-
-    override fun getKeys(): Set<String> = copywriter.getKeys()
-
-    override fun getDate(
-        date: LocalDateTime,
-        options: DateTimeFormatOptions,
-    ): String = copywriter.getDate(date, options)
-
-    override fun getAbridge(): String = copywriter.getAbridge()
 }
 
 class DesktopCopywriter(
